feat(hub-common): add function to set feed configuration
